requirements.txt
libscrc
pyyaml
pymodbus
paho-mqtt
